sleep phase updates
save a lot of energy by shutting of the rf part between sending. it is a interval of 2-3 seconds, so it saves
added gui to set program id
added a workaround for a very hard to fix bug when the battery is quite low in power. bug 1 sets it on
removed unused code