add information memory driver
Add code that manages and makes available the information memory. The
code is disabled even when selected with "make config" as long as it
is not actively used (modify project.h if you want to add your app to
the list that uses it).

Every app that wants to store data in the information memory should use
the provided functions to avoid any interference with other applications
that want to do the same.

You need to have mspgcc4>=20100829 when activated, otherwise an internal
compiler error will occur.