Avoid false peaks when re-enabling the altimeter.
The altimeter will turn itself off (to save battery) after some time,
if we reenable it after this or simply cycle through other functions
on line 1 while changing our altitude, we will see a false peak in
the vario displayand subsequently in the Vzmin and Vzmax stats.

This commit fixes this problem by ignoring the first seen altitude
difference.