Add flight time display and compile time selections
Here I add the flight time display, which can be reset by holding the
"#" button in flight time display mode.

I also added compile time flags in the source file to enable or disable
individual components in order to save space, as this addition makes the
code sufficiently large to require disabling some other feature to fit
in the watch. These flags are enabled by default, except for the display
of altitude change in Pascal.

One could for example disable the "test" feature which serves little purpose
for everyday use.